Created separate MVC for comments
