+++ app.js
@@ -5,7 +5,8 @@ app.use(express.json());
 
 const { getEndpoints } = require("./controllers/api.controller.js");
 const { getTopics } = require("./controllers/topics.controller.js");
-const { getArticles, getArticleById, getCommentsByArticleId, addCommmentToArticle } = require("./controllers/articles.controller.js");
+const { getArticles, getArticleById, getCommentsByArticleId } = require("./controllers/articles.controller.js");
+const { addCommmentToArticle } = require("./controllers/comments.controller.js");
 const { handleCustomErrors, handlePsqlErrors, handleServerErrors } = require("./errors/app.errors.js");
 
 app.get("/api", getEndpoints);
